+++ resources/newssite/news-next/src/data/index.js
