+++ resources/newssite/news-nuxt/data/index.js
